p4i integration Integrate 98028 from 3 to main: Shut up compiler.
[git-p4: depot-paths = "//depot/qt/main/": change = 98029]

+++ tools/assistant/profile.cpp
@@ -280,7 +280,7 @@ QString Profile::makeRelativePath( const QString &base, const QString &path )
     }
     QString rel = "";
     if ( bl.count() )
-	for ( int i = 0; i < bl.count(); ++i )
+	for ( int i = 0; i < (int)bl.count(); ++i )
 	    rel += "../";
     else
 	rel = "./";
